move clients back out of package
