+++ src/letsbuilda/pypi/__init__.py
@@ -1,6 +1,6 @@
 """A wrapper for PyPI's API and RSS feed."""
 
-from .clients.sync_client import PyPIServices
+from .sync_client import PyPIServices
 from .models import JSONPackageMetadata, RSSPackageMetadata
 
 __all__ = [
