+++ src/letsbuilda/pypi/async_client.py
@@ -11,7 +11,7 @@ except ImportError as error:
     msg = "Please install letsbuilda[async] for async support!"
     raise ImportError(msg) from error
 
-from letsbuilda.pypi.models import JSONPackageMetadata, RSSPackageMetadata
+from .models import JSONPackageMetadata, RSSPackageMetadata
 
 
 class PyPIServices:
